refactor: fix type mismatches in frequency/delay and simOutputSize APIs
- Change frequency(), delay(), setFrequency(), setDelay() from float
  to double — storage was already double, eliminating 12 lossy casts
- Change simOutputSize() from int to qsizetype to match QVector::size()

+++ MCP/Server/Handlers/ElementHandler.cpp
@@ -347,11 +347,11 @@ QJsonObject ElementHandler::handleSetElementProperties(const QJsonObject &params
     }
 
     if (params.contains("frequency") && element->hasFrequency()) {
-        float oldFreq = element->frequency();
-        float newFreq = static_cast<float>(params.value("frequency").toDouble());
+        double oldFreq = element->frequency();
+        double newFreq = params.value("frequency").toDouble();
 
-        oldProperties["frequency"] = static_cast<double>(oldFreq);
-        newProperties["frequency"] = static_cast<double>(newFreq);
+        oldProperties["frequency"] = oldFreq;
+        newProperties["frequency"] = newFreq;
 
         element->setFrequency(newFreq);
     }
@@ -373,7 +373,7 @@ QJsonObject ElementHandler::handleSetElementProperties(const QJsonObject &params
         oldProperties["delay"] = oldDelay;
         newProperties["delay"] = newDelay;
 
-        element->setDelay(static_cast<float>(newDelay));
+        element->setDelay(newDelay);
     }
 
     if (params.contains("trigger") && element->hasTrigger()) {
